new logic works, change next to all screens
