+++ App.js
@@ -2,105 +2,72 @@ import * as React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
 import { StatusBar } from 'expo-status-bar';
-import { StyleSheet, View } from 'react-native';
+import { Dimensions, StyleSheet, View } from 'react-native';
 import { useFonts } from 'expo-font';
-import Splash from './Screens/Splash';
-import Welcome from './Screens/Welcome';
-import Login from './Screens/Login';
-import CreateUser from './Screens/CreateUser';
-import ShiftScreen from './Screens/ShiftScreen';
-import ReportHours from './Screens/ReportHours';
-import ManageShifts from './Screens/ManageShifts';
-import DeleteShifts from './Screens/DeleteShifts';
-import MyShifts from './Screens/MyShifts';
-import History from './Screens/History';
-import EditEmails from './Screens/EditEmails';
-import UploadImgScreen from './Screens/UploadImgScreen';
-import Settings from './Screens/Settings';
+import TopBarComponent from './Components/TopBarComponent';
+import MainAppNavigator from './Components/MainAppNavigator';
+import AppEntryNavigator from './Components/AppEntryNavigator';
+import { useEffect } from 'react/cjs/react.development';
+import { GestureHandlerRootView } from 'react-native-gesture-handler';
+import BackgroundImage from './utility/BackGroundImage';
+
 
 const Stack = createStackNavigator();
 
 export default function App() {
-  const[hasLoggedIn , setHasLoggedIn] = React.useState(false);
+    const [hasLoggedIn, setHasLoggedIn] = React.useState(false); // triggeröi kumpi navigaattori käytössä returnissa
+    const [userData, setUserData] = React.useState(null); // käyttäjän tiedontallennukseen ja välittämiseen
 
-  const handleLogin = () => {
-    setHasLoggedIn(true);
-  };
 
-  const handleLogOut = () => {
-    setHasLoggedIn(false);
-  };
 
-  const [fontsLoaded] = useFonts({
-    'Saira-Regular': require('./assets/fonts/Saira-Regular.ttf'),
-  });
-  return (
-    <NavigationContainer>
-      <Stack.Navigator initialRouteName="Splash">
-        <Stack.Screen
-          name="Splash"
-          component={Splash}
-          options={{ headerShown: false }}
-        />
-        <Stack.Screen
-          name="Welcome"
-          component={Welcome}
-          options={{ headerShown: false }}
-        />
-        <Stack.Screen
-          name="Login"
-          component={Login}
-          options={{ headerShown: false }}
-        />
-        <Stack.Screen
-          name="CreateUser"
-          component={CreateUser}
-          options={{ headerShown: false }}
-        />
-        <Stack.Screen
-          name="ShiftScreen"
-          component={ShiftScreen}
-          options={{ headerShown: false }}
-        />
-        <Stack.Screen
-          name="ReportHours"
-          component={ReportHours}
-          options={{ headerShown: false }} 
-        />
-        <Stack.Screen
-          name="ManageShifts"
-          component={ManageShifts}
-          options={{ headerShown: false }}
-        />
-        <Stack.Screen
-          name="DeleteShifts"
-          component={DeleteShifts}
-          options={{ headerShown: false }}
-        />
-        <Stack.Screen
-          name="MyShifts"
-          component={MyShifts}
-          options={{ headerShown: false }}
-        />
-        <Stack.Screen
-          name="History"
-          component={History}
-          options={{ headerShown: false }}
-        />
-        <Stack.Screen
-          name="EditEmails"
-          component={EditEmails}
-          options={{ headerShown: false }}
-        />
-        <Stack.Screen
-          name="UploadImgScreen"
-          component={UploadImgScreen}
-          options={{ headerShown: false }}
-        />
-    
-          
-      </Stack.Navigator>
-      <StatusBar style="auto" />
-    </NavigationContainer>
-  );
+    const handleLogin = (loginResponse) => { 
+        setHasLoggedIn(true);
+        setUserData(loginResponse);
+    }
+
+    const handleLogOut = () => {
+        setHasLoggedIn(false);
+        setUserData(null);
+    }
+
+    useEffect(() => {   // ekstra vahti jotta tiedot ei varmasti jää muistiin
+        if(!hasLoggedIn) {
+            setUserData(null);
+        }
+    }, [hasLoggedIn]);
+
+    const [fontsLoaded] = useFonts({
+        'Saira-Regular': require('./assets/fonts/Saira-Regular.ttf'),
+    });
+
+    return ( 
+      
+      <NavigationContainer>
+        
+      
+        <View style={{ flex: 1} }>
+          {hasLoggedIn ? (
+              <>
+              
+              <GestureHandlerRootView>
+              <View style={{ height: screenHeight * 0.1}}> 
+                <TopBarComponent handleLogOut={handleLogOut} userRole={userData?.role} /> 
+              </View>
+              </GestureHandlerRootView>
+                <MainAppNavigator screenProps={{ userData: userData, handleLogout: handleLogOut}}/>
+              </>
+          ) : (
+              <AppEntryNavigator handleLogin={handleLogin} handleLogOut={handleLogOut} setUserData={setUserData} />
+          )}
+        </View>
+        <StatusBar style="auto" />
+      </NavigationContainer>
+  )
+  
 }
+
+const screenHeight = Dimensions.get('window').height;
+const screenWidth = Dimensions.get('window').width;
+const styles = StyleSheet.create({
+  
+});
