+++ Components/AppEntryNavigator.js
@@ -1,14 +1,18 @@
 import * as React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
-import Splash from './Screens/Splash';
-import Welcome from './Screens/Welcome';
-import Login from './Screens/Login';
-import CreateUser from './Screens/CreateUser';
+import Splash from '../Screens/Splash';
+import Welcome from '../Screens/Welcome';
+import Login from '../Screens/Login';
+import CreateUser from '../Screens/CreateUser';
+
+
 
 const Stack = createStackNavigator();
 
 export default function AppEntryNavigator({handleLogin, handleLogOut, setUserData}) {
+
+    
 
 
     return (
