+++ Components/Logout.js
@@ -12,7 +12,7 @@ const Logout = ({logOut}) => {
 
   const handleLogout = async () => {
     try {
-      let temporaryTokenSave = await getToken();
+      let temporaryTokenSave = await getToken(); // kopioidaan hetkeksi jotta saadaan myöhemmin servulle logout tehtyä kanssa -tero
       // Remove the token from storage
       await removeToken();
       // Check if the token is removed successfully
@@ -42,8 +42,8 @@ const Logout = ({logOut}) => {
 
       temporaryTokenSave = null; // tyhjennys varmuuden vuoksi -tero
 
-      // Navigate to the home screen
-      navigation.navigate('Welcome');
+      // Navigate to the home screen - triggeröi navigationcontainerin vaihdon -tero
+     logOut();
     } catch (error) {
       console.error('Error during logout:', error);
       Alert.alert('Logout Failed', 'An error occurred during logout.');
